update: create separate pages directory
